Tiny Blanbk gif to reduce file size
I was shocked when I saw the solution was to load a 119kb file in the head of the page (which blocks the browser from rending the page until it's loaded). That in itself has a negative SEO impact, as described here https://developers.google.com/speed/docs/insights/BlockingJS

This edit will considerably reduce the file size.